# Documentation/kernel/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cratis.Kernel.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/events/store/{microserviceId}/{tenantId}/sequence/{eventSequenceId}": {
      "post": {
        "tags": [
          "EventSequence"
        ],
        "summary": "Appends an event to the event log.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "description": "The microservice to append for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          },
          {
            "name": "eventSequenceId",
            "in": "path",
            "description": "The event sequence to append to.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventSequenceId"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "description": "The tenant to append to.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TenantId"
            }
          }
        ],
        "requestBody": {
          "description": "The payload with the details about the event to append.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendEvent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendEvent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AppendEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "EventSequence"
        ],
        "summary": "Get events for a specific event sequence in a microservice for a specific tenant.",
        "parameters": [
          {
            "name": "eventSequenceId",
            "in": "path",
            "description": "Event sequence to get for.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EventSequenceId"
            }
          },
          {
            "name": "microserviceId",
            "in": "query",
            "description": "Microservice to get for.",
            "schema": {
              "$ref": "#/components/schemas/MicroserviceId"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "description": "Tenant to get for.",
            "schema": {
              "$ref": "#/components/schemas/TenantId"
            }
          },
          {
            "name": "microserviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppendedEventWithJsonAsContent"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppendedEventWithJsonAsContent"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppendedEventWithJsonAsContent"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/events/store/{microserviceId}/{tenantId}/sequence/{eventSequenceId}/histogram": {
      "get": {
        "tags": [
          "EventSequence"
        ],
        "summary": "Get a histogram of a specific event sequence. PS: Not implemented yet.",
        "parameters": [
          {
            "name": "microserviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventSequenceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventHistogramEntry"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventHistogramEntry"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventHistogramEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppendEvent": {
        "type": "object",
        "properties": {
          "eventSourceId": {
            "$ref": "#/components/schemas/EventSourceId"
          },
          "eventType": {
            "$ref": "#/components/schemas/EventType"
          },
          "content": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonNode"
            },
            "description": "The content of the event represented as System.Text.Json.Nodes.JsonObject.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the payload for appending an event."
      },
...
      "AppendedEventWithJsonAsContent": {
        "type": "object",
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/EventMetadata"
          },
          "context": {
            "$ref": "#/components/schemas/EventContext"
          },
          "content": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "additionalProperties": false,
        "description": "Represents an event that has been appended to an event log with the content as JSON."
      },
      "CausationId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents an identifier for correlation."
      },
      "CausedBy": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents an identifier of an identity that was the root of a cause."
      },
...
      "CorrelationId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Represents an identifier for correlation."
      },
...
      "EventContext": {
        "type": "object",
        "properties": {
          "eventSourceId": {
            "$ref": "#/components/schemas/EventSourceId"
          },
          "sequenceNumber": {
            "$ref": "#/components/schemas/EventSequenceNumber"
          },
          "occurred": {
            "type": "string",
            "description": "<see cref=\"T:System.DateTimeOffset\">When</see> it occurred.",
            "format": "date-time"
          },
          "validFrom": {
            "type": "string",
            "description": "<see cref=\"T:System.DateTimeOffset\">When</see> event is considered valid from.",
            "format": "date-time"
          },
          "tenantId": {
            "$ref": "#/components/schemas/TenantId"
          },
          "correlationId": {
            "$ref": "#/components/schemas/CorrelationId"
          },
          "causationId": {
            "$ref": "#/components/schemas/CausationId"
          },
          "causedBy": {
            "$ref": "#/components/schemas/CausedBy"
          },
          "observationState": {
            "$ref": "#/components/schemas/EventObservationState"
          }
        },
        "additionalProperties": false,
        "description": "Represents the context in which an event exists - typically what it was appended with."
      },
      "EventGeneration": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Represents the generation of an Cratis.Events.EventType."
      },
      "EventHistogramEntry": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "EventMetadata": {
        "type": "object",
        "properties": {
          "sequenceNumber": {
            "$ref": "#/components/schemas/EventSequenceNumber"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          }
        },
        "additionalProperties": false,
        "description": "Represents the metadata related to an event."
      },
      "EventObservationState": {
        "enum": [
          0,
          1,
          2,
          4,
          8
        ],
        "type": "integer",
        "description": "Represents the observation state for an event.",
        "format": "int32"
      },
      "EventSequenceId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          },
          "isEventLog": {
            "type": "boolean",
            "description": "Get whether or not this is the default log event sequence.",
            "readOnly": true
          },
          "isOutbox": {
            "type": "boolean",
            "description": "Get whether or not this is the default outbox event sequence.",
            "readOnly": true
          },
          "isInbox": {
            "type": "boolean",
            "description": "Get whether or not this is the default outbox event sequence.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the unique identifier of an event sequence."
      },
...
      "EventSequenceNumber": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "description": "Represents the sequence number within an event log for an event."
      },
      "EventSourceId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "nullable": true
          },
          "isSpecified": {
            "type": "boolean",
            "description": "Check whether or not the Cratis.Events.EventSourceId is specified.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "Represents the unique identifier of an instance of an event source."
      },
      "EventType": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EventTypeId"
          },
          "generation": {
            "$ref": "#/components/schemas/EventGeneration"
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether or not the event type is considered a public event."
          }
        },
        "additionalProperties": false,
        "description": "Represents the type of an event."
      },
      "EventTypeId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents the concept of the unique identifier of a type of event."
      },
...
      "JsonNode": {
        "type": "object",
        "properties": {
          "options": {
            "$ref": "#/components/schemas/JsonNodeOptions"
          },
          "parent": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "root": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "additionalProperties": false
      },
      "JsonNodeOptions": {
        "type": "object",
        "properties": {
          "propertyNameCaseInsensitive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "MicroserviceId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents the concept of the microservice identifier."
      },
...
      "TenantId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "Gets or inits the underlying value for the instance.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Represents the unique identifier of a tenant in the system."
      },
...
    }
  }
}